+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 <a name="Pending Release"></a>
 ## [Pending Release](https://github.com/lightstep/otel-launcher-java/compare/0.8.1...master)
+* Replaced lightstep.hostname tag with host.name
 
 <a name="0.8.1"></a>
 ## [0.8.1](https://github.com/lightstep/otel-launcher-java/compare/0.8.0...0.8.1)
-* Set lighstep.hostname tag
+* Set lightstep.hostname tag
 * Fixed setting log level 
 
 <a name="0.8.0"></a>